# server/docs/swagger.yaml
host: 127.0.0.1:1323
info:
  contact:
    name: API Support
  description: WISMED Wisite-API Server. Updated@ 2022-09-03T11:33:31+10:00
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: WISMED WISITE API
  version: "1.0"
paths:
...
  /api/rel/action/{whom}:
    put:
      consumes:
      - application/json
      parameters:
      - description: which action to apply, accept [follow, unfollow, block, unblock,
          mute, unmute]
        in: query
        name: action
        required: true
        type: string
      - description: whose uname you want to follow
        in: path
        name: whom
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - following successfully
        "400":
          description: Fail - invalid action type
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: relation actions
      tags:
      - Relation
  /api/rel/content/{type}:
    get:
      consumes:
      - application/json
      parameters:
      - description: relation content type to apply, accept [following, follower,
          blocked, muted]
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - got following successfully
        "400":
          description: Fail - invalid relation content type
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: get all relation users for one type
      tags:
      - Relation
...
  /api/sign/in:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: user name or email
        in: formData
        name: uname
        required: true
        type: string
      - description: password
        format: password
        in: formData
        name: pwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - sign-in successfully
        "400":
          description: Fail - incorrect password
        "500":
          description: Fail - internal error
      summary: sign in action. if ok, got token
      tags:
      - Sign
  /api/sign/new:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: unique user name
        in: formData
        name: uname
        required: true
        type: string
      - description: user's email
        format: email
        in: formData
        name: email
        required: true
        type: string
      - description: user's real full name
        in: formData
        name: name
        required: true
        type: string
      - description: user's password
        in: formData
        name: pwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - then waiting for verification code
        "400":
          description: Fail - invalid registry fields
        "500":
          description: Fail - internal error
      summary: sign up action, step 1. send user's basic info for registry
      tags:
      - Sign
  /api/sign/reset-pwd:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: unique user name
        in: formData
        name: uname
        required: true
        type: string
      - description: user's email
        format: email
        in: formData
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - then waiting for verification code
        "400":
          description: Fail - invalid registry fields
        "500":
          description: Fail - internal error
      summary: reset password action, step 1. send verification code to user's email
        for authentication
      tags:
      - Sign
  /api/sign/verify-email:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: unique user name
        in: formData
        name: uname
        required: true
        type: string
      - description: verification code (in user's email)
        in: formData
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - sign-up successfully
        "400":
          description: Fail - incorrect verification code
        "500":
          description: Fail - internal error
      summary: sign up action, step 2. send back email verification code
      tags:
      - Sign
  /api/sign/verify-reset-pwd:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: unique user name
        in: formData
        name: uname
        required: true
        type: string
      - description: verification code (in user's email)
        in: formData
        name: code
        required: true
        type: string
      - description: new password
        in: formData
        name: pwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK   - password updated successfully
        "400":
          description: Fail - incorrect verification code
        "500":
          description: Fail - internal error
      summary: reset password action, step 2. send back verification code for updating
        password
      tags:
      - Sign
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: authorization
    type: apiKey
swagger: "2.0"